update: beat: {Double <= Int}

+++ src/main/scala/com/github/nnnnusui/musicdsl/input/Sound.scala
@@ -2,5 +2,5 @@ package com.github.nnnnusui.musicdsl.input
 
 sealed trait Sound
 object Sound {
-  case class Get(sampleRate: Int, tempo: Double, beat: Int)
+  case class Get(sampleRate: Int, tempo: Double, beat: Double)
 }
